feat: Update quantity handling across various controllers to support decimal values
- Changed validation rules for quantity fields in PurchaseClaimController, PurchaseController, SaleController, SaleItemController, SaleReturnController, and StockController to accept numeric values with a minimum of 0.001.
- Updated internal handling of quantities to use float instead of int for better precision in calculations.
- Added new invoice template enum and updated printer settings migration to include new fields for invoice templates and footer lines.
- Converted quantity columns in relevant tables from INTEGER to DECIMAL to allow for fractional quantities.
- Enhanced inventory valuation services to create stock records with initial average costs if they do not exist.
- Updated seeder to create a larger number of customers, vendors, and products per branch for testing purposes.

+++ app/Services/InventoryValuationWriteService.php
@@ -9,7 +9,7 @@ class InventoryValuationWriteService
     /**
      * Receives purchase into product_stocks and updates avg_cost (moving average).
      */
-    public function receivePurchase(int $productId, ?int $branchId, int $receiveQty, float $unitPrice, ?string $ref = null): void
+    public function receivePurchase(int $productId, ?int $branchId, float $receiveQty, float $unitPrice, ?string $ref = null): void
     {
         if ($receiveQty <= 0) return;
 
@@ -19,7 +19,7 @@ class InventoryValuationWriteService
             ->where('branch_id',  $branchId)
             ->first();
 
-        $oldQty  = (int)($row->quantity ?? 0);
+        $oldQty  = (float)($row->quantity ?? 0);
         $oldCost = (float)($row->avg_cost ?? 0.0);
 
         $newQty  = $oldQty + $receiveQty;
@@ -50,7 +50,7 @@ class InventoryValuationWriteService
     public function returnToVendor(
         int $productId,
         int $branchId,
-        int $returnQty,
+        float $returnQty,
         ?string $ref = null
     ): float {
         // Expect positive qty for a return. If 0 or negative, just echo current avg and do nothing.
@@ -81,7 +81,7 @@ class InventoryValuationWriteService
         }
 
         $avg    = (float) ($row->avg_cost ?? 0.0);
-        $onHand = (int)   ($row->quantity ?? 0);
+        $onHand = (float) ($row->quantity ?? 0);
 
         // NO CLAMPING: allow going negative
         $newQty = $onHand - $returnQty;
